requirements/requirements.txt
#pandas==1.5.2
#prefect==2.7.7
prefect
#prefect-sqlalchemy==0.2.2
prefect-gcp[cloud_storage]==0.2.4
protobuf==4.21.11
pyarrow==10.0.1
pandas-gbq==0.18.1
psycopg2-binary==2.9.5
sqlalchemy==1.4.46
#wget
kaggle
gcsfs